pyproject.toml
[build-system]
requires = ["setuptools >= 45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "textual-zptess"
dynamic = ["version"]
description = "A package to collect measurements published by TESS instruments into a SQlite database"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Rafael González", email = "rafael08@ucm.es"},
]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
dependencies = [
   'pypubsub',   # Publish/Subscribe support Model/View/Controller
   'SQLAlchemy[asyncio]',
   'aiosqlite',
   'anyio',
   'httpx',
   'textual-dev'
]

[project.optional-dependencies]
test = [
    "pytest>=6",
]

[project.urls]
Homepage = "https://github.com/astrorafael/textual-zptess"
Repository = "https://github.com/astrorafael/textual-zptess.git"

[project.scripts]
zptess = "zptess.main:main"
schema = "zptess.dbase.schema:main"
phot = "zptess.photometer.service:main"

# pyproject.toml
# Automatically generates a _version-py file from with Git information
# This file shold be placed in .gitignore
[tool.setuptools_scm]
write_to = "src/zptess/_version.py"

[tool.setuptools.packages.find]
# scanning for namespace packages is true by default in pyproject.toml, so
# you do NOT need to include the following line.
namespaces = true
where = ["src"]

[tool.setuptools]
script-files = [
#    "scripts/tessdb_flush",
]

# Needed for Textual CSS files
[tool.setuptools.package-data]
"zptess.tui.css" = ["*.css"]
